tests: Move hijack test to dedicated folder
